==> notebooks/step00_utils.ipynb <==
---
title: Step 00 - Utilities
subject: Churn Analysis
subtitle: Step 00 - Utilities - Churn Analysis
short_title: Utilities
date: 2025-12-17

affiliations:
  - id: "ucb"
    name: "University of California, Berkeley"

authors:
  - name: Jocelyn Perez
    affiliations: ["ucb"]
    email: jocelyneperez@berkeley.edu
    orcid: 0009-0009-0231-9254

  - name: Claire Kaoru Shimazaki
    affiliations: ["ucb"]
    email: ckshimazaki@berkeley.edu
    orcid: 0009-0001-0828-3370

  - name: Colby Zhang
    affiliations: ["ucb"]
    email: colbyzhang@berkeley.edu
    orcid: 0009-0005-4786-6922

  - name: Olorundamilola Kazeem
    affiliations: ["ucb"]
    email: dami@berkeley.edu
    orcid: 0000-0003-2118-2221

# https://mystmd.org/guide/frontmatter#frontmatter-downloads
# https://mystmd.org/guide/website-downloads
# downloads:
#   -  ...

# https://mystmd.org/guide/website-downloads#include-exported-pdf
# exports:
#   - format: pdf
#     template: lapreprint-typst
#     output: exports/my-document.pdf
#     id: my-document-export
# downloads:
#   - id: my-document-export
#     title: A PDF of this document

exports:
  - format: pdf
    template: lapreprint-typst
    output: ../pdf_builds/step00_utils/step00_utils_ipynb_to.pdf
    line_numbers: true

license: BSD-3-Clause

keywords: utilities, churn, spotify

abstract: What are the utilities?
---


# Step 00: Utilities

import src.step00_utils as step00_utils

step00_utils.DIR_PROJECT_CURRENT

step00_utils.DIR_PROJECT_HOME

step00_utils.DIR_DATA
